requirements.txt
beautifulsoup4==4.12.3
pandas==2.2.2
python-dotenv==1.0.1
Requests==2.32.3
selenium==4.23.1
black==23.3.0
isort==5.10.1
clean-text==0.6.0
unidecode==1.3.8
psycopg2==2.9.9
sqlalchemy==2.0.34